@@ -1,23 +1,23 @@
-import { ApplicationConfig, importProvidersFrom } from '@angular/core';
+import { ApplicationConfig } from '@angular/core';
 import { provideRouter, withViewTransitions } from '@angular/router';
 
 import { routes } from './app.routes';
-import { transition } from '@angular/animations';
-import { HttpClientModule } from '@angular/common/http';
+import {
+  provideHttpClient,
+  withInterceptorsFromDi,
+} from '@angular/common/http';
 
 export const appConfig: ApplicationConfig = {
-  //Agregar transisiones al cambiar de ruta con view transitions
   providers: [
     provideRouter(
       routes,
       withViewTransitions({
         skipInitialTransition: true,
         onViewTransitionCreated(transitionInfo) {
+          console.log(transitionInfo);
         },
-      }),
+      })
     ),
-    importProvidersFrom(
-      HttpClientModule
-    )
-  ]
+    provideHttpClient(withInterceptorsFromDi()),
+  ],
 };
