@@ -1,11 +1,42 @@
-import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  computed,
+  signal,
+} from '@angular/core';
+import { TitleComponent } from '@shared/title/title.component';
 
 @Component({
   standalone: true,
-  imports: [],
+  imports: [TitleComponent, CommonModule],
   templateUrl: './change-detection.component.html',
-  styles: ``
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  styles: ``,
 })
 export default class ChangeDetectionComponent {
+  public currentFramekork = computed(() => {
+    return `Change Detection - ${this.frameworkAsSignal().name}`
+  })
 
+  public frameworkAsSignal = signal({
+    name: 'Angular',
+    releaseDate: 2016,
+  });
+
+  public frameworkAsProperty = {
+    name: 'Angular',
+    releaseDate: 2016,
+  };
+
+  constructor() {
+    setTimeout(() => {
+      // this.frameworkAsProperty.name = 'React';
+      this.frameworkAsSignal.update((value) => ({
+        ...value,
+        name: 'React',
+      }));
+      console.log('Hecho');
+    }, 3000);
+  }
 }
