@@ -1,56 +1,42 @@
 import { CommonModule } from '@angular/common';
 import { HttpClient } from '@angular/common/http';
-import { Component, computed, inject, signal } from '@angular/core';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  computed,
+  inject,
+  signal,
+} from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { TitleComponent } from '@shared/title/title.component';
 import { User } from '../../../interfaces/req-response';
-import { toSignal } from '@angular/core/rxjs-interop'
+import { toSignal } from '@angular/core/rxjs-interop';
 import { switchMap } from 'rxjs';
 import { UsersService } from '@services/users.service';
 
 @Component({
   standalone: true,
+  changeDetection: ChangeDetectionStrategy.OnPush,
   imports: [CommonModule, TitleComponent],
-  template: `
-    <app-title title="User" />
-
-    @if(user()) {
-      <section>
-        <img [srcset]="user()!.avatar" [alt]="user()!.first_name" />
-
-        <div>
-          <h3>{{titleLabel()}}</h3>
-          <p>{{user()!.email}}</p>
-        </div>
-      </section>
-    }@else {
-      <p>Cargando Info</p>
-    }
-  `,
-  styles: ``
+  templateUrl: './user.component.html',
 })
 export default class UserComponent {
-
   private route = inject(ActivatedRoute);
   private usersService = inject(UsersService);
 
-
-  titleLabel = computed(() => {
-    if (this.user()) {
-      return `Información del usuario ${this.user()?.first_name} ${this.user()?.last_name}`
-    }
-    return 'Información del usuario';
-  })
-
-  // public user = signal<User | undefined>(undefined)
   public user = toSignal(
     this.route.params.pipe(
       switchMap(({ id }) => this.usersService.getUserById(id))
     )
-  )
+  );
+  
 
-  //  constructor(){
-  //   this.route.params.
-  //  }
-
+  titleLabel = computed(() => {
+    if (this.user()) {
+      return `Información del usuario ${this.user()?.first_name} ${
+        this.user()?.last_name
+      }`;
+    }
+    return 'Información del usuario';
+  });
 }
