@@ -1,7 +1,7 @@
 import { Injectable, computed, inject, signal } from '@angular/core';
-import { User, UserResponse, UsersResponse } from '../interfaces/req-response';
 import { HttpClient } from '@angular/common/http';
 import { delay, map } from 'rxjs';
+import { User, UserResponse, UsersResponse } from '@interfaces/req-response';
 
 interface State {
   users: User[];
@@ -14,7 +14,8 @@ interface State {
 export class UsersService {
   private http = inject(HttpClient)
 
-  #state = signal<State>({
+  
+   #state = signal<State>({
     loading: true,
     users: [],
   })
@@ -24,10 +25,9 @@ export class UsersService {
 
   constructor() {
 
-    this.http.get<UsersResponse>('https://reqres.in/api/users')
+    this.http.get<UsersResponse>('/reqres-api/api/users')
       .pipe( delay(1500) )
       .subscribe( res => {
-
         this.#state.set({
           loading: false,
           users: res.data,
@@ -37,8 +37,8 @@ export class UsersService {
 
   }
 
-  getUserById(id: string) {
-    return this.http.get<UserResponse>(`http://reqres.in/api/users/${id}`).pipe(
+  getUserById(id: number) {
+    return this.http.get<UserResponse>(`/reqres-api/api/users/${id}`).pipe(
       delay(1500),
       map(res => res.data)
     )
