@@ -5,7 +5,7 @@ import { Component, Input } from '@angular/core';
   selector: 'app-heavy-loaders-slow',
   standalone: true,
   imports: [CommonModule],
-  template: `<section [ngClass]="['w-full h-[600px]', cssClass]">Hola Mundo</section>`,
+  template: `<section [ngClass]="['w-full h-[600px]', cssClass]">Heavy Loader Slow</section>`,
   styles: ``
 })
 export class HeavyLoadersSlowComponent {
@@ -16,7 +16,7 @@ export class HeavyLoadersSlowComponent {
     const start = Date.now()
     while (Date.now() - start < 3000) {
     }
-    console.log('cargando')
+    console.log('cargado')
 
   }
 }
