@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, TrackByFunction } from '@angular/core';
 import { routes } from '../../app.routes';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
@@ -8,18 +8,13 @@ import { RouterModule } from '@angular/router';
   standalone: true,
   imports: [CommonModule, RouterModule],
   templateUrl: './sidemenu.component.html',
-  styleUrl: './sidemenu.component.scss'
+  styleUrl: './sidemenu.component.scss',
 })
 export class SidemenuComponent {
-
   public menuItems = routes
-    .map(route => route.children ?? [])
+    .map((route) => route.children ?? [])
     .flat()
-    .filter(route => route && route.path)
-    .filter(route => !route.path?.includes(':'));
-  constructor() {
-
-  }
-
-  
+    .filter((route) => route && route.path)
+    .filter((route) => !route.path?.includes(':'));
+  constructor() {}
 }
